introduce RESTC_CPP_FETCH_DEPENDENCIES
CMake's ExternalProject fetcher is handy but only useful when no
other package manager for dependencies is present. If we want to use
vcpkg, Yocto or similar facilities which can provide dependencies as
packages, we need to turn off CMake's fetcher.

The default behaviour is still the same that CMake will fetch RapidJSON
and Lest. Nevertheless it can now be turned off optionally.